Refactor daily session dialogs to reset form state by lifecycle instead of $effect
- split add/modify drawer form logic into mount-on-open content components
- remove reset-on-open effects from daily drawers and delete dialog
- keep daily component validation clean under Svelte 5 rules
